Fix for PAVE broken world model
PAVE has an issue with one of its world models showing up as missing (red exclamation point) in the game.  This is due to the name of the file being different to what it is being pointed at in the ESP.  Renaming the NIF file fixes the issue.  I propose to add these special instructions until the PAVE gets an update that fixes the problem.

Special Instructions:  There is an issue with the TTW added T-45d Power Armor Of Love.  One of the models name is misspelled causing the world object model to show as a missing object (a red exclemation mark model will show up in its place).  Renaming the model will fix the problem.  Here is now to fix it.

- Double click on PAVE and select the Filetree tab. Now expand `meshes\armor\PAVE\T45PAPink`.  Rename `OutfitGO.nif` to `OutfitArmyGO.nif`.